requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
jinja2==3.1.2
pydantic[email]==2.5.0